Working Commit Before Major Change In The Parsing System
This is not functional, but working code. However, since JsonObject and
JsonArray parsing may cause problems if both classes are enclosed in
each other since currently special characters are only detected inside
JSON formatted strings, the parsing mechanic will be reworked to an
event based system.
This commit is only for reverting purposes.